% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.gradLasso}
\alias{coef.gradLasso}
\title{Extract Model Coefficients}
\usage{
\method{coef}{gradLasso}(object, ...)
}
\arguments{
\item{object}{A gradLasso fitted object.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of coefficients.
}
\description{
Extract Model Coefficients
}
