% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.gradLasso}
\alias{fitted.gradLasso}
\title{Extract Fitted Values}
\usage{
\method{fitted}{gradLasso}(object, ...)
}
\arguments{
\item{object}{A gradLasso fitted object.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of fitted values.
}
\description{
Extract Fitted Values
}
