% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{grad_binomial}
\alias{grad_binomial}
\title{Binomial Family (Logistic Regression)}
\usage{
grad_binomial()
}
\value{
A list containing gradient, deviance, and prediction functions for logistic regression.
}
\description{
Binomial Family (Logistic Regression)
}
