% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{grad_gaussian}
\alias{grad_gaussian}
\title{Gaussian Family (Least Squares)}
\usage{
grad_gaussian()
}
\value{
A list containing gradient, deviance, and prediction functions for Gaussian regression.
}
\description{
Gaussian Family (Least Squares)
}
