% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{grad_negbin}
\alias{grad_negbin}
\title{Negative Binomial Family}
\usage{
grad_negbin()
}
\value{
A list containing gradient, deviance, and prediction functions for Negative Binomial regression.
}
\description{
Negative Binomial Family
}
