% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/families.R
\name{grad_zinb}
\alias{grad_zinb}
\title{Zero-Inflated Negative Binomial Family}
\usage{
grad_zinb()
}
\value{
A list containing gradient, deviance, and prediction functions for ZINB regression.
}
\description{
Zero-Inflated Negative Binomial Family
}
