% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.gradLasso}
\alias{logLik.gradLasso}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{gradLasso}(object, ...)
}
\arguments{
\item{object}{A gradLasso fitted object.}

\item{...}{Additional arguments.}
}
\value{
An object of class \code{logLik}.
}
\description{
Extract Log-Likelihood
}
