% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.gradLasso}
\alias{print.summary.gradLasso}
\title{Print method for summary}
\usage{
\method{print}{summary.gradLasso}(x, ...)
}
\arguments{
\item{x}{A summary.gradLasso object.}

\item{...}{Additional arguments passed to print.}
}
\value{
Invisibly returns the input object.
}
\description{
Print method for summary
}
