% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.gradLasso}
\alias{residuals.gradLasso}
\title{Extract Residuals}
\usage{
\method{residuals}{gradLasso}(object, ...)
}
\arguments{
\item{object}{A gradLasso fitted object.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of residuals.
}
\description{
Extract Residuals
}
