% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradLasso.R
\name{gradLasso}
\alias{gradLasso}
\title{Gradient Descent LASSO with Stability Selection}
\usage{
gradLasso(
  formula,
  data = NULL,
  family = grad_gaussian(),
  lambda = NULL,
  lambda_cv = TRUE,
  standardize = TRUE,
  cv_subsample = NULL,
  parallel = FALSE,
  n_cores = NULL,
  boot = TRUE,
  n_boot = 50,
  boot_ci = c(0.025, 0.975),
  batch_size = NULL,
  warm_start = TRUE
)
}
\arguments{
\item{formula}{Formula object. Supports pipes for ZINB (e.g., y ~ x1 + x2 | z1).}

\item{data}{Data frame.}

\item{family}{Family object.}

\item{lambda}{Optional fixed lambda.}

\item{lambda_cv}{Configuration for CV.}

\item{standardize}{Logical. Standardize predictors?}

\item{cv_subsample}{Integer. Speedup for CV.}

\item{parallel}{Logical. Enable parallel processing?}

\item{n_cores}{Integer. Number of cores.}

\item{boot}{Logical. Run stability selection?}

\item{n_boot}{Number of bootstraps.}

\item{boot_ci}{Vector of two probabilities for CIs.}

\item{batch_size}{Integer. Mini-batch SGD.}

\item{warm_start}{Logical. Warm start bootstraps.}
}
\description{
Gradient Descent LASSO with Stability Selection
}
